# runtime: Optimize FindSymbol