--- a/pkg/runtime/elf_symbols.go
+++ b/pkg/runtime/elf_symbols.go
@@ -14,8 +14,10 @@
 package runtime
 
 import (
+	"bufio"
 	"bytes"
 	"debug/elf"
+	"encoding/binary"
 	"errors"
 	"fmt"
 	"io"
@@ -57,19 +59,18 @@ func IsSymbolNameInDynamicSymbols(f *elf.File, matches [][]byte) (bool, error) {
 	return isSymbolNameInSection(f, elf.SHT_DYNSYM, matches)
 }
 
-func isSymbolNameInSection(f *elf.File, t elf.SectionType, matches [][]byte) (bool, error) {
-	symtabSection := f.SectionByType(t)
+func isSymbolNameInSection(ef *elf.File, t elf.SectionType, matches [][]byte) (bool, error) {
+	symtabSection := ef.SectionByType(t)
 	if symtabSection == nil {
 		return false, elf.ErrNoSymbols
 	}
 
-	if symtabSection.Link <= 0 || symtabSection.Link >= uint32(len(f.Sections)) {
-		return false, errors.New("section has invalid string table link")
+	strtabReader, err := stringTableReader(ef, symtabSection.Link)
+	if err != nil {
+		return false, fmt.Errorf("cannot load string table section: %w", err)
 	}
 
-	s := f.Sections[symtabSection.Link].Open()
-
-	sr, err := ainur.NewStreamReader(s, 8192)
+	sr, err := ainur.NewStreamReader(strtabReader, 8192)
 	if err != nil {
 		return false, fmt.Errorf("create stream reader: %w", err)
 	}
@@ -94,31 +95,240 @@ func isSymbolNameInSection(f *elf.File, t elf.SectionType, matches [][]byte) (bo
 	return false, nil
 }
 
+func scanNullTerminated(data []byte, atEOF bool) (int, []byte, error) {
+	if atEOF && len(data) == 0 {
+		return 0, nil, nil // io.EOF
+	}
+
+	if i := bytes.IndexByte(data, 0); i >= 0 {
+		// We have a full null-terminated line.
+		return i + 1, data[0:i], nil
+	}
+
+	// If we're at EOF, we have a final, non-terminated line. Return it.
+	if atEOF {
+		return len(data), data, nil
+	}
+
+	// Request more data.
+	return 0, nil, nil
+}
+
+func firstIndexOfMatchingSymbol(r io.Reader, matches [][]byte) (int, error) {
+	scanner := bufio.NewScanner(r)
+	scanner.Split(scanNullTerminated)
+
+	bytesRead := 0
+	for scanner.Scan() {
+		for _, match := range matches {
+			b := scanner.Bytes()
+			if len(b) == len(match) && bytes.Equal(b, match) {
+				return bytesRead, nil
+			}
+		}
+		bytesRead += len(scanner.Bytes()) + 1 // +1 for null terminator
+	}
+	if err := scanner.Err(); err != nil {
+		return -1, fmt.Errorf("scanner: %w", err)
+	}
+	return -1, nil
+}
+
 // FindSymbol finds symbol by name in the given elf file.
 func FindSymbol(ef *elf.File, symbol string) (*elf.Symbol, error) {
-	symbols, err := ef.Symbols()
+	sym, err := getSymbol(ef, elf.SHT_SYMTAB, symbol)
 	// If there are no symbols, try dynamic symbols.
 	if err != nil && !errors.Is(err, elf.ErrNoSymbols) {
-		return nil, fmt.Errorf("error reading ELF symbols: %w", err)
+		return nil, fmt.Errorf("error getting ELF symbols: %w", err)
+	}
+	if sym != nil {
+		return sym, nil
 	}
 
-	for _, s := range symbols {
-		if s.Name == symbol {
-			return &s, nil
-		}
-	}
-
-	dynamicSymbols, err := ef.DynamicSymbols()
+	sym, err = getSymbol(ef, elf.SHT_DYNSYM, symbol)
 	if err != nil {
 		return nil, fmt.Errorf("error reading ELF dynamic symbols: %w", err)
 	}
-	for _, s := range dynamicSymbols {
-		if s.Name == symbol {
-			return &s, nil
-		}
+	if sym != nil {
+		return sym, nil
 	}
 
 	return nil, fmt.Errorf("symbol %q not found", symbol)
+}
+
+func getSymbol(ef *elf.File, typ elf.SectionType, symbol string) (*elf.Symbol, error) {
+	switch ef.Class {
+	case elf.ELFCLASS64:
+		return getSymbol64(ef, typ, symbol)
+
+	case elf.ELFCLASS32:
+		return getSymbol32(ef, typ, symbol)
+
+	case elf.ELFCLASSNONE:
+		fallthrough
+
+	default:
+		return nil, fmt.Errorf("unknown ELF class: %v", ef.Class)
+	}
+}
+
+func getSymbol32(ef *elf.File, typ elf.SectionType, symbol string) (*elf.Symbol, error) {
+	symtabSection := ef.SectionByType(typ)
+	if symtabSection == nil {
+		return nil, elf.ErrNoSymbols
+	}
+
+	strdataReader, err := stringTableReader(ef, symtabSection.Link)
+	if err != nil {
+		return nil, fmt.Errorf("cannot load string table section: %w", err)
+	}
+
+	match, err := firstIndexOfMatchingSymbol(strdataReader, [][]byte{[]byte(symbol)})
+	if err != nil {
+		return nil, fmt.Errorf("cannot find symbol: %w", err)
+	}
+	if match == -1 {
+		return nil, fmt.Errorf("symbol not found, %q", symbol)
+	}
+
+	data, err := symtabSection.Data()
+	if err != nil {
+		return nil, fmt.Errorf("cannot load symbol section: %w", err)
+	}
+
+	symtab := bytes.NewReader(data)
+	if symtab.Len()%elf.Sym64Size != 0 {
+		return nil, errors.New("length of symbol section is not a multiple of Sym64Size")
+	}
+
+	// The first entry is all zeros.
+	var skip [elf.Sym32Size]byte
+	_, err = symtab.Read(skip[:])
+	if err != nil {
+		return nil, fmt.Errorf("cannot read first entry: %w", err)
+	}
+
+	var sym elf.Sym32
+	for symtab.Len() > 0 {
+		if err := binary.Read(symtab, ef.ByteOrder, &sym); err != nil {
+			return nil, fmt.Errorf("cannot read symbol: %w", err)
+		}
+
+		if sym.Name != uint32(match) {
+			continue
+		}
+
+		str, err := readStringFromReaderSeeker(strdataReader, int(sym.Name))
+		if err != nil {
+			return nil, fmt.Errorf("cannot read symbol name: %w", err)
+		}
+
+		return &elf.Symbol{
+			Name:    str,
+			Info:    sym.Info,
+			Other:   sym.Other,
+			Section: elf.SectionIndex(sym.Shndx),
+			Value:   uint64(sym.Value),
+			Size:    uint64(sym.Size),
+		}, nil
+	}
+
+	return nil, fmt.Errorf("symbol not found, %q", symbol)
+}
+
+func getSymbol64(ef *elf.File, typ elf.SectionType, symbol string) (*elf.Symbol, error) {
+	symtabSection := ef.SectionByType(typ)
+	if symtabSection == nil {
+		return nil, elf.ErrNoSymbols
+	}
+
+	strdataReader, err := stringTableReader(ef, symtabSection.Link)
+	if err != nil {
+		return nil, fmt.Errorf("cannot load string table section: %w", err)
+	}
+
+	match, err := firstIndexOfMatchingSymbol(strdataReader, [][]byte{[]byte(symbol)})
+	if err != nil {
+		return nil, fmt.Errorf("cannot find symbol: %w", err)
+	}
+	if match == -1 {
+		return nil, fmt.Errorf("symbol not found, %q", symbol)
+	}
+
+	data, err := symtabSection.Data()
+	if err != nil {
+		return nil, fmt.Errorf("cannot load symbol section: %w", err)
+	}
+
+	symtab := bytes.NewReader(data)
+	if symtab.Len()%elf.Sym64Size != 0 {
+		return nil, errors.New("length of symbol section is not a multiple of Sym64Size")
+	}
+
+	// The first entry is all zeros.
+	var skip [elf.Sym64Size]byte
+	_, err = symtab.Read(skip[:])
+	if err != nil {
+		return nil, fmt.Errorf("cannot read first entry: %w", err)
+	}
+
+	var sym elf.Sym64
+	for symtab.Len() > 0 {
+		if err := binary.Read(symtab, ef.ByteOrder, &sym); err != nil {
+			return nil, fmt.Errorf("cannot read symbol: %w", err)
+		}
+
+		if sym.Name != uint32(match) {
+			continue
+		}
+
+		str, err := readStringFromReaderSeeker(strdataReader, int(sym.Name))
+		if err != nil {
+			return nil, fmt.Errorf("cannot read symbol name: %w", err)
+		}
+
+		return &elf.Symbol{
+			Name:    str,
+			Info:    sym.Info,
+			Other:   sym.Other,
+			Section: elf.SectionIndex(sym.Shndx),
+			Value:   sym.Value,
+			Size:    sym.Size,
+		}, nil
+	}
+
+	return nil, fmt.Errorf("symbol not found, %q", symbol)
+}
+
+// readStringFromReaderSeeker extracts a zero-terminated string from an ELF string table io.Reader.
+func readStringFromReaderSeeker(rs io.ReadSeeker, start int) (string, error) {
+	_, err := rs.Seek(int64(start), io.SeekStart)
+	if err != nil {
+		return "", err
+	}
+
+	var buf bytes.Buffer
+	for {
+		var b [1]byte
+		if _, err := rs.Read(b[:]); err != nil {
+			return "", err
+		}
+		if b[0] == 0 {
+			break
+		}
+		buf.Write(b[:])
+	}
+
+	return buf.String(), nil
+}
+
+// stringTable reads and returns the string table given by the
+// specified link value.
+func stringTableReader(ef *elf.File, link uint32) (io.ReadSeeker, error) {
+	if link <= 0 || link >= uint32(len(ef.Sections)) {
+		return nil, errors.New("section has invalid string table link")
+	}
+	return ef.Sections[link].Open(), nil
 }
 
 // ReadStringAtAddress reads a null-terminated string from the given address in